feat(stripe): check if stripe keys are for live environment (#91)
* feat(stripe): check if stripe keys are for live environment

* fix(registration): import statemeht
