+++ src/controller/payment.controller.ts
@@ -15,11 +15,14 @@ const stripe = require('stripe')(process.env.STRIPE_SECRET_KEY);
  * Controller for handling payments.
  */
 export default class PaymentCtrl extends BaseController {
-  stripeService = new StripeManager();
-  serviceService = new ServiceManager();
+  stripeService: StripeManager;
+  serviceService: ServiceManager;
 
   constructor() {
     super(enrollments);
+
+    this.stripeService = new StripeManager();
+    this.serviceService = new ServiceManager();
   }
 
   /**
@@ -123,68 +126,7 @@ export default class PaymentCtrl extends BaseController {
     }
   }
 
-  checkForStripe = async (req: Request, res: Response) => {
-    const enabled_events = [
-      'invoice.payment_succeeded',
-      'invoice.payment_failed',
-      'customer.subscription.deleted',
-      'customer.subscription.updated'
-    ];
-    try {
-      const isHttps = (process.env.API_HOST || '').includes('https://');
-      let webhooks = false,
-        webhookEvents = false,
-        newEndpointCreated = false,
-        webhookEndpoint: any,
-        account: any;
-      if
-        (process.env.STRIPE_SECRET_KEY &&
-        process.env.STRIPE_PUBLIC_KEY &&
-        process.env.STRIPE_ENROLLMENT_SECRET
-      ) {
-        account = await stripe.account.retrieve();
-        const endpoints = await stripe.webhookEndpoints.list();
-        webhookEndpoint = endpoints?.data?.find((endpoint: any) => endpoint.url === `${process.env.API_HOST}/webhooks`);
-
-        if (isHttps && !webhookEndpoint && !process.env.STRIPE_WEBHOOKS_KEY) {
-          webhookEndpoint = (await stripe.webhookEndpoints.create({
-            enabled_events,
-            url: `${process.env.API_HOST}/webhooks`,
-          }));
-
-          if (webhookEndpoint) newEndpointCreated = true;
-        }
-
-        if (!!webhookEndpoint) webhooks = true;
-        if (_isEqual(webhookEndpoint?.enabled_events, enabled_events)) webhookEvents = true;
-      }
-
-      return res
-        .status(200)
-        .json({
-          isHttps,
-          stripeKey: !!process.env.STRIPE_SECRET_KEY ? true : false,
-          stripePublicKey: !!process.env.STRIPE_PUBLIC_KEY ? true : false,
-          enrollmentSecret: !!process.env.STRIPE_ENROLLMENT_SECRET ? true : false,
-          stripeWebhooksKey: !!process.env.STRIPE_WEBHOOKS_KEY ? true : false,
-          newEndpointCreated,
-          webhooks,
-          webhookEvents,
-          rnUrl: process.env.REQUEST_NETWORK_UI_URL,
-          account: {
-            requirements: {
-              currently_due: account?.requirements?.currently_due || [],
-              eventually_due: account?.requirements?.eventually_due || [],
-              past_due: account?.requirements?.past_due || []
-            },
-            capabilities: account?.capabilities || {}
-          }
-        });
-    } catch (error: Error | unknown) {
-      Logger.error("Error creating token:" + JSON.stringify(error));
-      return res
-        .status(500)
-        .json({ ok: false, error: (error as Error)?.message || "Internal server error" });
-    }
+  checkForStripe = async (req?: Request, res?: Response) => {
+    return this.stripeService.checkForStripe(req, res);
   }
 }
