+++ src/core/registration.ts
@@ -3,6 +3,7 @@ import Logger from "../utils/logger";
 import { AuthenticatedRequest, XTaoshiHeaderKeyType } from "./auth-request";
 import ServiceManager from "../service/service.manager";
 import { ServiceDTO } from "../db/dto/service.dto";
+import StripeManager from "../service/stripe.manager";
 
 /**
  * Handles registration of the API instance with the UI application.
@@ -25,10 +26,12 @@ export default class Registration {
       )
     );
 
+    const stripeStatus = await new StripeManager().checkForStripe();
     const body = {
       baseApiUrl: AuthenticatedRequest.baseURL,
       apiPrefix: process.env.API_PREFIX,
       validators,
+      stripeStatus
     };
 
     try {
